lots and lots of stuff

+++ src/rndr.h
@@ -1,7 +1,7 @@
 #ifndef RNDR_H__
 #define RNDR_H__
 
-extern uint8 *pixels;
+extern uint32 *pixels;
 
 #define FOV 3.141593 / 2.0
